Added new config option
