+++ assets/config.json.tmpl
@@ -1,6 +1,7 @@
 {
  "askHost": "{{ .Env.ASK_URL }}",
  "recaptcha": "{{ .Env.RECAPTCHA }}",
+ "published_base_url": "{{ .Env.PUBLISHED_BASE_URL }}",
   "s3": {
     "bucket": "{{ .Env.S3_BUCKET }}",
     "region": "{{ .Env.AWS_REGION }}",
